refactor(rule): move stateless processor helpers into the helpers file
The buildAddressComparisonString, findRecyclerAccreditation, and
pickGpsComment functions were defined as loose top-level functions
below the processor class. They have no dependency on processor state,
so they belong in geolocation-and-address-precision.helpers.ts
alongside the other stateless helpers in this rule. Switches the
declarations from 'function' to 'const' arrow form to match the
existing helpers file convention.
